Queries --> Schedule
Fixes #4 Fixes #2.
Queries did not make a lot of sense as a resource name. Schedule makes
more sense because we are searching for schedules, we want to look at
a schedule, etc. We are making queries to view a schedule, so it might
be apt to name some class that handles the querying part that.