WIP: Term.
This is a work in progress commit which introduces the Term object. This
obj will be responsible for fetching the current term.

I am having some issues testing it, rails doesn't like it that I am not
inheriting from ActiveRecord and so it keeps throwing errors. Perhaps I
need to disable ActiveRecord entirely.