Add Subject.
We need to store the subjects that are available. We are storing this
since we need to retrieve the course schedules via the
`term/subject/schedules` endpoint and also for future features which
will require us to list out the subjects for filtering purposes. This
commit introduces a Subject model with two attributes:

* code:string
* description:string

This is all we need for now, validations and relations will be added as
required.