Introduce Subject#update.
Fixes #42.
The update method overrides the one provided by ActiveRecord since we
want custom functionality here. This is not a standard CRUD resource as
a result, we will never a use case where we need to create a new subject
via the UI, this will be done internally at all times. Thus we can
safely override the update method to fetch a list of all subjects that
are taught at the University and save each of them. If a subject already
exists, then the method does nothing.
An obvious side effect is what if a subject is no longer taught? This
subject will always remain in the db. For our purposes, we will be
fetching the schedules based on the subjects and courses being
taught in the current term so we won't run into a `record not found`
issue. However, we should introduce a method that handles the cleanup
some time later.